Add files via upload
This application is very useful for any organisations or the students to secure their data aswell to generate the password and many other options along with that many generators are available in this and you can find the QR code and can also generator the QR code. using the public and private can protect the data. And keys will update for every 5-10secs from the hardware entropy.